KPMP-720 renamed dynamicForm~ to DynamicForm~, as it is a helper class
